Always Shutdown PG Before pgBackRest Repo Host
Ensures PG instances are always shutdown prior to shutting down the
pgBackRest repo host, such as when "Shutdown" is set to "true" in the
PostgresCluster spec or when shutting down the cluster gracefully prior
to a major PG upgrade.  This ensures the pgBackRest dedicated repository
host is available for the full lifetime of any/all running PG instances,
and WAL can therefore always be pushed as needed.

Additionally, when setting "Shutdown" to "true" in the PostgresCluster
spec in order to gracefully shutdown the cluster  prior to a major PG
upgrade, the "PostgresVersion" is now also set to the
"FromPostgresVersion".  This ensures the cluster continues to reconcile
with the proper PG version (i.e. while the cluster is shutting down)
prior to the upgrade completing.

Issue: [sc-13291]