Do Not Remove Certain Status for Major PG Upgrade
When preparing a cluster for a major PG upgrade, only the "repos"
section of the pgBackRest status is removed, while the
"PostgresDataInitialized" condition (if present) is now also preserved.
This ensures that only the status and conditions that are pertinent to
the upgrade process are removed, and all other status is preserved.

Issue: [sc-13310]