Allow passwords to be generated without symbols
PostgreSQL does not limit passwords to a particular character set nor
encoding. To stay reasonably compatible with most client libraries and
tools, we generate passwords in the US-ASCII character set and encoding
which is also UTF-8.

However, getting passwords to tools often involves encodings and formats
in which ASCII symbols have special meaning and need to be escaped:
ampersand U+0026 in YAML, brackets U+003C and U+003E in XML, question
mark U+003F in URIs, and so on. Anything that neglects to escape these
on the path between the Kubernetes Secret and the tool renders the tool
inoperative.

This commit adds a "spec.users.password" field to PostgresCluster that
allows password generation to be configured per PostgreSQL user. The
only option so far is the ability to choose which characters may appear
in the password. The default and current behavior is called "ASCII".
A new choice, called "AlphaNumeric", produces passwords without any
ASCII symbols.

Issue: [sc-12862]
Issue: CrunchyData/postgres-operator#2781
Issue: CrunchyData/postgres-operator#2935