Move pgBackRest Logs Out of the /tmp Directory
Currently, the pgBackRest logs stored in the PostgreSQL instance
and pgBackRest repo Pods are located in the /tmp directory. This
is limited to the default 16MB emptyDir configuration set by PGO
and is not sufficient in certain cases.

This update changes the log locations. PGBackRest logs from the
PostgreSQL instance will now be stored at '/pgdata/pgbackrest/log'
on the pgData volume. PGBackRest logs from the repo host will be
stored at '/pgbackrest/repo#/log' where 'repo#' corresponds to the
first configured repository on the PostgresCluster spec that is
configured with a storage volume. That is, if the first configured
repo is 'repo1', and 'repo1' has a configured volume,  the log path
will be '/pgbackrest/repo1/log' so that those logs will be stored on
the 'repo1' volume. However, if 'repo1' does not have a configured
volume, the path will be set based so that the logs are stored on the
first repo configured with a storage volume.

Please note that the initial backup is taken on the first configured
repository. As such, this will determine where certain initial logs
are stored, i.e. db-backup.log, etc

Issue: [sc-13395]