E2E Test for GSSAPI Authentication With PG
Adds an end-to-end test to verify the ability to properly configure
PostgreSQL for GSSAPI authentication.  This is specifically done
using Kerberos, i.e. a ticket issued by a Kerberos KDC server
is utilized to log into PostgreSQL when running the test.

Please note that this specific test has been added to a "e2e-other"
directory, since additional configuration and infrastructure (e.g. a
a running Kerberos Key Distribution Center) must be in place in order
to successfully run the test.

[sc-13319]