Disable errchkjson linter temporarily
The "errchkjson" linter looks specifically at calls that encode or
marshal values to JSON. It does not ignore assignments to the blank
identifier like the "errcheck" linter. Instead, it recommends ignoring
the error only when the type being marshaled is guaranteed to marshal.

Without configuration, it recommends changes to our tests. When
configured to analyze the types being marshaled, it recommends changes
to our code. Disable it entirely for now and add it to the list of
possible future linters.

Issue: [sc-13607]