Generate a coverage report during GitHub actions
This uses GitHub artifacts and the reports provided by the built-in
"go tool cover" command. Functions without complete coverage are printed
to the GitHub action logs, and a full HTML report is uploaded as an
artifact. The total percent of statements covered appears on the
workflow run summary as an annotation.