Copy pgBackRest configuration resources when performing a restore
This commit copies pgBackRest configuration and related ConfigMaps
and Secrets used by the source cluster as VolumeProjections when
bootstrapping a new cluster using pgBackRest restore. The
configuration copied are given a unique name to avoids conflicts
with any other clusters and is removed once cluster bootstrap is
complete. If the ConfigMap or Secret VolumeProjections are configured
as "optional" in the source cluster, those resources may be absent
without any error or failure when the copying is attempted.

Issue: [sc-13601]