Move pgAdmin Startup Script to PGO
Right now when pgAdmin is deployed by PGO, scripts within the
container are utilized as the entrypoint and command for running
the container.

This commit converts the startup script to a commands within the
pgAdmin Pod template to lessen the dependency on scripts contained
within the pgAdmin container. Additionally, this commit updates the
NSS_WRAPPER setup to use the standard functions used by other Pods
managed by PGO.

Note that the various configuration files within the container
(whether for Apache, pgAdmin, etc.) are still be utilized to
configure pgAdmin.

Issue: [sc-13629]