Operator logging for database init SQL failures
If there is an error in the init SQL that runs as part of reconcileDatabaseInitSQL, then there is no way for the user to know what the error is.  Adding this additional log statement will make it easier for users to know when init sql operations have succeeded and/or failed.  It also brings this part of the code up to par with other similar operations in the codebase.

Issue: #3029