Enable Support for Secrets for pgAdmin LDAP_BIND_PASSWORD
This commit allows the use of a mounted Secret that contains
the desired LDAP_BIND_PASSWORD setting for pgAdmin which is used
for the pgAdmin login user when LDAP authentication is enabled.
This allows the password value to be encoded in a Secret rather
than stored in plaintext in a ConfigMap with the other pgAdmin
configuration settings.

Issue: [sc-13712]