Disable the "gci" Go linter
Recent changes to this linter trigger false positives in a majority of
files. It was enabled as part of the "format" preset. We did nothing to
configure it so its behavior matches that of "goimports". The latter is
also enabled as part of the "format" preset and supports the "--fix"
flag to automatically correct imports before committing.