Kuttl test to create a cluster and resize the PVC
This test creates two simple clusters with a single primary and a repo
host. In the first cluster we create data then increase the size of the
pvc. Then we check that the pvc size has changed, the size matches the
new expected side and the data is still present.

In the second cluster we attempt decrease the size of the volume and
expect the PersistentVolumeError.