Create an EventRecorder for tests
The FakeRecorder provided by "k8s.io/client-go" coverts each event to a
string and sends it to a channel. Tests that want to check the Type or
Reason separately from the Message have to resort to regexp captures.

Instead, this implementation does the same work as EventRecorder without
trying to batch or correlate isomorphic events. Calls to the recorder
are stored in a slice of events/v1.Event that tests can interrogate.