Change the volume claim test into a unit test
This is the last API test that requires a full Kubernetes cluster, and
it flakes during PR checks. We added an end-to-end KUTTL test for
resizing volumes in 112c910df, so the remaining value of this test is
in contrived scenarios that trigger a handful of error paths.

Reduce the test to those paths with errors mimicking those from the API.
Describe the scenarios that lead to those errors and link to their
origins in Kubernetes.

Issue: [sc-14270]
See: 112c910df14a793a56b106d3b3ed409f7bfd512f
Co-authored-by: jmckulk <joseph.mckulka@crunchydata.com>