Quarantine flaky delete test
The `delete` test that looked at event timestamps to make sure the
replica deleted before the primary occasionally flaked out.
This PR removes that timestamp checking, quarantining that version
of the test for future debugging; and changes the in-use test to
simply verify that a cluster with replica deletes.

Issue [sc-15009]