Allow NodePort Port to be Specified via the PostgresCluster Spec
This update allows a specific NodePort port to be specified for
the primary Postgres, pgBouncer and pgAdmin services via the
PostgresCluster spec.

Note this is used when type is NodePort or LoadBalancer only.
Setting this value when using the 'ClusterIP' type will result
in an error. The specified value must be also be in-range and
not currently in use or the operation will fail. If unspecified,
a port will be allocated if this Service requires one as before.

Resolves #3008
Issue: [sc-14918]