Align psql Job Backoff Limit & Restart Policy
Sets the "backoffLimit" to "6", and the "restartPolicy" to "never",
for all psql Jobs in the Kuttl test suite.  This has been done to
to address psql Jobs that are sometimes reaching the current backoff
limit and failing, while also better aligning all psql Jobs
within the Kuttl test suite.  Additionally, a "restartPolicy" of
"never" should also help facilitate the debugging of failed psql Jobs.