Remove the postmaster.pid file prior to pgBackRest restore
This commit removes the postmaster.pid file, if it exists, from the PGDATA
directory before attempting a restore. This allows the restore to be tried
more than once without causing an error due to the presence of the file
in subsequent attempts or in scenarios where the file is otherwise present.

Issue: [sc-15157]