Update Standby Replication Diagrams
This commit updates the existing repo-based standby cluster
configuration diagram to a new Draw.io generated image and adds
the associated XML file. It also creates two new diagrams to
illustrate a streaming standby cluster configuration and a cluster
that configured to have both a streaming standby and an external repo.

Issue: [sc-14710]