Custom TLS for Exporter (Encryption Only)
With this change we allow users to bring custom certificates and enable
TLS for the exporter. This will be an opt-in feature, PGO will not
automatically generate certs like it does for some other features.

You can enable TLS by using the following spec fields:

spec:
  monitoring:
    pgmonitor:
      exporter:
        customTLSSecret:
          name: hippo.tls

Once TLS is enabled in the exporter, you can configure your Prometheus
instance to scrape over https.