Update pgBackRest repo option logic
When taking a backup, PGO tries to help by not allowing the user
to pass the "--repo" option. However, the current method for catching
this results in catching any option that begins with "--repo", which
prevents users from passing in perfectly valid options.

This commit corrects the flag check to only block on exact matches of
"--repo".

Issue: [sc-16128]