Remove disable exporter tls test
Checking that tls has been disabled on a cluster (where it was
previously enabled) is difficult. This is because we need to wait for
the instance pod to be redeployed without tls configuration. We are
removing case from the kuttl test with plans to ensure we have the same
coverage in go tests in the future.

Issue: [sc-16572]