Allow to control the order of shared_preload_libraries (specifically for citus extension)
The citus extension needs to be loaded first when bringing up a cluster. The current implementation allows customising the extensions loaded, but extensions provided by the resource definition will be loaded after any internal ones.

Therefore, we updated the implementation such that the extension list order will be driven by the shared_preload_libraries list when present.

When your definition has a shared_preload_libraries list, the extension order will change to put the shared_preload_libraries first. ***This may change the behaviour of existing definitions when the order relied on the internal libs list***.

When there shared_preload_libraries list contains one or more or the internal extensions, the duplicate internal extensions will be ignored, which allows the user to fully define the order of loading of extensions.

Issue: #3194