Exporter tests overwrite annotation
As part of testing, kuttl will add annotations to instances pods before
and after a change. Kuttl tests will continue to loop through the script
until various conditions are met. This means that the annotation may be
run more than once. If this happens, kubectl will complain that the
annotation is already set and needs to be overwritten. This change adds
the overwrite flag to kubectl annotate commands.