Tweak restart logic for exporter (#3817)
* Tweak restart logic for exporter

This separates out the kill/restart logic for the exporter;
previously, if a file changed, we would kill/restart. This
led to some test flakes where the restart would happen too
quickly (hypothesis) before the port was free. This PR separates
the logic:

If the watched files change, kill the exporter;
If no exporter is running, start the exporter.

This also adds a check to the start_postgres_exporter func:
save the PID to a file only if that proc is running.

Issue: [PGO-420]