Updates for an always-on pgBackRest repo host
To support the 'backup-standby' pgBackRest configuration setting
as well as to simplify the overall pgBackRest integration, this
commit changes the pgBackRest repo host creation behavior to always
create a 'repo host' Pod as a location to run commands, regardless
of whether or not a repo volume is defined. This allows backup
commands to be consistently run on this Pod instead of being run at
times on the primary instance Pod.

Note that in cases where a repo host volume is not defined in the
PostgresCluster spec, no volume will be created and pgBackRest log
files will not be available in the Pod.

Issue: PGO-562