Adds the 'environment' Field to the PostgresCluster Spec
An 'environment' field is now available within the PostgresCluster spec,
which allows the user to specify the infrastructure and/or stage in
development the PostgresCluster is associated with.  Accepted values for
this field are 'production' and 'development'. 'production' is the default.

Additionally, a CEL validation rule has been added to the CRD that requires
backups to be configured for 'production' PostgresCluster's.  This change is
fully backward compatible and non-breaking with existing PostgresCluster
specs.

Issue: PGO-1645