Update Controller Gen to v0.16.3
Updates to the latest controller-gen release.  CRD's and RBAC have been
regenerated, and "namespace" has been removed from the markers in the
Patroni and pgBackRest rbac.go files (it was no longer providing much benefit
since the go code already cleanly organizes the RBAC, and changes to controller
controller-gen had the potential to break RBAC generation as a result of its use).