Replace our CRD validation patch with CEL rules
The JSON patch was awkward to maintain, and we forgot to update it when
we added PVCs to our APIs.

I considered defining these rules on a shared Go type in our API
package, but I did not like the type conversion it requires in our
controller and test code.

Issue: PGO-1748