Consistently use either pointer or value receivers
Subtle bugs can arise when mixing the two. A type is more coherent when
it behaves as only one of (1) a value or (2) a reference.

A new linter identifies this situation but does not yet account for
unmarshal methods on value types.

See: https://go.dev/wiki/CodeReviewComments#receiver-type
See: https://go.dev/wiki/MethodSets