Configure Patroni Logs to be stored in a file
This commit updates the Patroni configuration so that the instance
Pod logs go to a file on the 'pgdata' volume instead of to stdout.
This file is located at /pgdata/patroni/log/patroni.log.

This is accomplished by setting 'log.dir' in the Patroni YAML
configuration. Setting 'log.file_size' to 0 means that the log
file will not rollover.

- https://patroni.readthedocs.io/en/latest/yaml_configuration.html#log
- https://github.com/patroni/patroni/blob/v3.3.4/patroni/log.py#L431-L432
- https://docs.python.org/3.11/library/logging.handlers.html#rotatingfilehandler

Issue: PGO-1701