Reconcile a Service Account for the pgBackRest Repo Host
Currently, the pgBackRest repo host uses the 'default' service
account. However, EKS's IAM role integration requires a specific
annotation to enable this feature. This change adds a new SA
for the repo host to allow PGO to reconcile a SA with this annotation,
thus allowing the IAM integration to work as expected.

fixes #4006
Issue: PGO-2123