Simplify upgrade scripts with more variables
I find this Go code and resulting Bash easier to read. This also logs
more about what is happening without changing the sequence of commands.

The script included an unnecessary `|| exit`, so I moved the `set -eu`
out of the Bash invocation into the script itself to make that behavior
more obvious.