Modify context path of OTEL transformers (#4134)
* Modify context path of OTEL transformers

As of 0.119.0, the parser collector modifies certain
paths of transforms. While this is not an error, the
resulting log can look a little alarming, so we rewrote
some of our transforms to match.

* Handle pgBouncer 1.24.0

During testing, some pgbouncer metrics were seen to cause
errors when accessing PG through pgbouncer; in addition,
we found that 1.24.0 introduced a change to the `show databases`
columns. This PR addresses those errors and that change.

NOTE: Even when not using NULL values, sqlqueryreceiver will
still warn when it finds a NULL value in some row that it scans.

Issue: PGO-2268
See: 7a006019b63199fbd88161c0dce67a13669556ff