Add pg_hba checksum metric (#4169)
* Add pg_hba checksum metric

We want OTel metrics to be as close to parity with
the postgres_exporter solution as possible; so this
PR adds the initial setup and metric query to get
'ccp_pg_hba_checksum', as well as the util function
to mark a pg_hba setting as valid.

Note: when adding some comments, some changes
were introduced to these functions as they exist in pgMonitor,
but the behavior is equivalent.

Issues: [PGO-2395]