Remove Divisor from Downward API resrouceFieldRefs
This means CPU and memory information will now be exposed using the default
divisor for CPU and memory (with default being "1" for both).  This means memory
information will now be represented in bytes, as expected by pgMonitor and the
CPK Metrics & Monitoring stack when  consuming Downward API information.

Issue: PGO-2604