Add additional volumes to pgbouncer (#4240)
Add additional volume for pgbouncer pods

* Add additional volume for pgbouncer pods

Change the API to allow users to specify preexisting
PVCs to attach to specified containers in the pgbouncer
instance pods.

The spec allows users to specify whether to add
the volume to
- all containers (by omitting the containers list)
- no containers (by specifying an empty containers list)
- a list of named containers

If any of the named containers isn't present, we continue
to reconcile, but issue a warning event with the names
of the missing containers.

* Move additional volumes struct to shared_types
* Don't need v1 of volumes struct

Issues: [PGO-2557]