Handle user-set pgBouncer logfile for OTEL, mkdir (#4278)
* Handle user-set pgBouncer logfile for OTEL

pgBouncer allows users to set the log destination; when OTEL
is enabled, we assume that destination is the default that we
set. This PR patches that hole by setting OTEL to use the
either the default or the user-set log destination.

This PR also has the pgBouncer mkdir the appropriate directory
if given that configuration.

Issues: [PGO-2565]