Add validation for pgbouncer logfile (#4280)
* Add validation for pgbouncer logfile

This PR adds validation for v1 pgBouncer so that the logfile
must either be in /tmp or one of the attached additional volumes.

This PR also adds validation for v1beta1 that restricts the number
of global config settings and requires a '.log' suffix on the
logfile.

This PR also fixes an error in the OTEL setup around a custom
logfile.

Issues: [PGO-2565]