Run Shellcheck on instance reload command
Shellcheck was being run on the reload command for pgBackRest. During
the auto-grow work, we noticed that the reload command for the instance
was not being checked. Since we had to update the auto-grow bash logic to
pass Shellcheck for pgBackRest, it seemed appropriate to enable the checks for the
instance as well.